Refactored Footer, Header, GlobalModal, SplashScreen
