+++ app/src/containers/Footer/index.js
@@ -1,8 +1,1 @@
-import React from 'react';
-import packageJson from '../../../package.json';
-
-import './Footer.css';
-
-export default () => (
-  <div className="footer">v{packageJson.version}</div>
-);
+export * from './Footer';
