+++ app/src/containers/SplashScreen/SplashScreen.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import './SplashScreen.css';
+
+export const SplashScreen = () => (
+  <div className="splash-screen">
+    <div className="splash-screen__logo" />
+  </div>
+);
