+++ app/src/containers/SplashScreen/index.js
@@ -1,11 +1,1 @@
-import React from 'react';
-
-import './SplashScreen.css';
-
-const SplashScreen = () => (
-  <div className="splash-screen">
-    <div className="splash-screen__logo" />
-  </div>
-);
-
-export default SplashScreen;
+export * from './SplashScreen';
